+++ scripts/macros/2014/classFeatures/bard/collegeOfSwords/bladeFlourish.js
@@ -1,5 +1,4 @@
-import {activityUtils, combatUtils, compendiumUtils, constants, dialogUtils, effectUtils, errors, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../../utils.js';
-
+import {activityUtils, combatUtils, dialogUtils, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../../utils.js';
 async function attack({trigger: {entity: item}, workflow}) {
     if (['boomingBlade', 'greenFlameBlade'].includes(genericUtils.getIdentifier(workflow.item))) return;
     if (!workflowUtils.isAttackType(workflow, 'weaponAttack')) return;
@@ -39,7 +38,7 @@ async function damage({trigger: {entity: item}, workflow}) {
     if (skipUses) bardicDie = {formula: '1d6'};
     if (!bardicDie) return;
     let damageType = workflow.defaultDamageType;
-    await workflowUtils.bonusDamage(workflow, bardicDie.formula, {damageType});
+    await workflowUtils.bonusDamage(workflow, bardicDie.formula, {damageType, ignoreCrit: itemUtils.getConfig(item, 'canCrit')});
     let rollResult = workflow.damageRolls.at(-1).total;
     switch (selection) {
         case 'DF':
@@ -93,5 +92,15 @@ export let bladeFlourish = {
                 priority: 50
             }
         ]
-    }
+    },
+    config: [
+        {
+            value: 'canCrit',
+            label: 'CHRISPREMADES.Config.CanCrit',
+            type: 'checkbox',
+            default: false,
+            category: 'homebrew',
+            homebrew: true
+        }
+    ]
 };